Add AgentConnect Provider
Keycloak does not support ES256 signature with OIDC provider, so we need to verify by ourselves.
AgentConnect publishes the keys on a JWKS URL but for the RSA Key, there is not usage, so we cannot directly
use the lkeycloak way and need to internalise it to do some null checking.

Signed-off-by: Cédric Couralet <cedric.couralet@insee.fr>